requirements.txt
langchain
unstructured
chromadb
tiktoken
openai
python-dotenv